w3c-reports/12092014/mac/runned-tests/IndexedDB/idbtransaction-expected.txt

CONSOLE MESSAGE: line 13: TypeError: null is not an object (evaluating 'indexedDB.open')

Harness Error. harness_status.status = 1 , harness_status.message = TypeError: null is not an object (evaluating 'indexedDB.open')
